requirements.txt
2to3==1.0
alabaster==0.7.13
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.12.0
beautifulsoup4==4.11.2
binaryornot==0.4.4
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
clr==1.0.3
colorama==0.4.6
constantly==15.1.0
cookiecutter==2.1.1
coverage==7.2.1
cryptography==39.0.1
daphne==3.0.2
Django==3.2.14
django-browser-reload==1.7.0
django-nose==1.4.7
django-tailwind==3.5.0
djangorestframework==3.14.0
docutils==0.19
factory-boy==3.2.1
Faker==16.8.1
hyperlink==21.0.0
idna==3.4
imagesize==1.4.1
incremental==22.10.0
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.2
msgpack==1.0.4
nose==1.3.7
nose-py3==1.5.0
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
pinocchio==0.4.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
service-identity==21.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.4
Sphinx==6.1.3
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
text-unidecode==1.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zope.interface==5.5.2